fix mocks directory
