--- packages/modules/pendle/src/actions/__mock__/account.ts
@@ -1,1 +1,0 @@
-export const ANVIL_PRIVATE_KEY = '0xac0974bec39a17e36ba4a6b4d238ff944bacb478cbed5efcae784d7bf4f2ff80'
